requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers
psycopg2-binary
Fernet
cryptography
django_environ==0.11.2
django-storages==1.14.2
httplib2==0.22.0
Pillow==9.0.0
PyJWT
psycopg2-binary